changed framework support to enums
